Avoid runner container (#127)

+++ test/prepare-10-reset.sh
@@ -1,16 +1,23 @@
+#!/usr/bin/env bash
+
 set -e
 
-echo "Reset for a fresh test ... "
+title="$1"
+schema="$2"
+verified="$3"
+repository="$4"
+
+echo "Reset for a fresh test - ${title:?} ... "
 
 helm repo ls -o yaml | yq -r '.[].name' | xargs -I {} helm repo rm {}
 
-rm -rf /schema/* &>/dev/null || true
-rm -rf /templates/* &>/dev/null || true
-rm -rf /repository &>/dev/null || true
-rm -rf /verified-schema &>/dev/null || true
+rm -rf "${schema:?}/*" &>/dev/null || true
+rm -rf "${repository:?}/*" &>/dev/null || true
+rm -rf "${verified:?}/*" &>/dev/null || true
 
-mkdir -p /verified-schema
-mkdir -p /repository/git
+mkdir -p "$verified"
+mkdir -p "$repository/git"
+mkdir -p "$repository/http"
 
 echo "done"
 echo
